backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.31.1
SQLAlchemy==2.0.36
alembic==1.13.2
pydantic==2.9.2
pydantic-settings==2.6.1
psycopg2-binary==2.9.9
python-multipart==0.0.17
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
httpx==0.27.2
